Realizada a implementação da política de acesso aos grupos de ingredientes.

+++ app/Http/Controllers/IngredientGroupsController.php
@@ -14,22 +14,26 @@ class IngredientGroupsController extends Controller
 
     public function store(Request $request)
     {
+        $this->authorize('create', IngredientGroup::class);
         return IngredientGroup::create($request->all());
     }
 
     public function show(IngredientGroup $ingredient_group)
     {
+        $this->authorize('view', $ingredient_group);
         return $ingredient_group;
     }
 
     public function update(Request $request, IngredientGroup $ingredient_group)
     {
+        $this->authorize('update', $ingredient_group);
         $ingredient_group->update($request->all());
         return $ingredient_group;
     }
 
     public function destroy(IngredientGroup $ingredient_group)
     {
+        $this->authorize('delete', $ingredient_group);
         $ingredient_group->delete();
         return $ingredient_group;
     }
